+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/PortletPreferenceDataFactory.java
@@ -1,0 +1,343 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+import com.liferay.asset.kernel.model.AssetCategoryModel;
+import com.liferay.asset.kernel.model.AssetTagModel;
+import com.liferay.blogs.constants.BlogsPortletKeys;
+import com.liferay.dynamic.data.lists.constants.DDLPortletKeys;
+import com.liferay.dynamic.data.lists.model.DDLRecordSetModel;
+import com.liferay.dynamic.data.mapping.constants.DDMPortletKeys;
+import com.liferay.journal.constants.JournalPortletKeys;
+import com.liferay.journal.model.JournalArticleResourceModel;
+import com.liferay.portal.kernel.model.PortletConstants;
+import com.liferay.portal.kernel.model.PortletPreferencesModel;
+import com.liferay.portal.kernel.portlet.PortletIdCodec;
+import com.liferay.portal.kernel.portlet.PortletPreferencesFactory;
+import com.liferay.portal.kernel.util.ObjectValuePair;
+import com.liferay.portal.kernel.util.PortletKeys;
+import com.liferay.portal.model.impl.PortletPreferencesModelImpl;
+import com.liferay.portlet.PortletPreferencesFactoryImpl;
+import com.liferay.portlet.PortletPreferencesImpl;
+import com.liferay.wiki.constants.WikiPortletKeys;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javax.portlet.PortletPreferences;
+
+/**
+ * @author Lily Chi
+ */
+public class PortletPreferenceDataFactory extends BaseDataFactory {
+
+	public static PortletPreferenceDataFactory getInstance() {
+		return _portletPreferenceDataFactory;
+	}
+
+	public String getPortletId(String portletPrefix) {
+		return portletPrefix.concat(PortletIdCodec.generateInstanceId());
+	}
+
+	public PortletPreferencesFactory getPortletPreferencesFactory() {
+		return _portletPreferencesFactory;
+	}
+
+	public List<PortletPreferencesModel>
+		newAssetPublisherPortletPreferencesModels(long plid) {
+
+		List<PortletPreferencesModel> portletPreferencesModels =
+			new ArrayList<>(3);
+
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, BlogsPortletKeys.BLOGS,
+				PortletConstants.DEFAULT_PREFERENCES));
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, JournalPortletKeys.JOURNAL,
+				PortletConstants.DEFAULT_PREFERENCES));
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, WikiPortletKeys.WIKI,
+				PortletConstants.DEFAULT_PREFERENCES));
+
+		return portletPreferencesModels;
+	}
+
+	public List<PortletPreferencesModel> newDDLPortletPreferencesModels(
+		long plid) {
+
+		List<PortletPreferencesModel> portletPreferencesModels =
+			new ArrayList<>(3);
+
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, DDLPortletKeys.DYNAMIC_DATA_LISTS_DISPLAY,
+				PortletConstants.DEFAULT_PREFERENCES));
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, DDLPortletKeys.DYNAMIC_DATA_LISTS,
+				PortletConstants.DEFAULT_PREFERENCES));
+		portletPreferencesModels.add(
+			_newPortletPreferencesModel(
+				plid, DDMPortletKeys.DYNAMIC_DATA_MAPPING,
+				PortletConstants.DEFAULT_PREFERENCES));
+
+		return portletPreferencesModels;
+	}
+
+	public List<PortletPreferencesModel> newJournalPortletPreferencesModels(
+		long plid) {
+
+		return Collections.singletonList(
+			_newPortletPreferencesModel(
+				plid, JournalPortletKeys.JOURNAL,
+				PortletConstants.DEFAULT_PREFERENCES));
+	}
+
+	public PortletPreferencesModel newPortletPreferencesModel(
+			long plid, long groupId, String portletId, int currentIndex,
+			long[] assetClassNameIds)
+		throws Exception {
+
+		if (currentIndex == 1) {
+			return _newPortletPreferencesModel(
+				plid, portletId, PortletConstants.DEFAULT_PREFERENCES);
+		}
+
+		String assetPublisherQueryName = "assetCategories";
+
+		if ((currentIndex % 2) == 0) {
+			assetPublisherQueryName = "assetTags";
+		}
+
+		ObjectValuePair<String[], Integer> objectValuePair = null;
+
+		Integer startIndex = _assetPublisherQueryStartIndexes.get(groupId);
+
+		if (startIndex == null) {
+			startIndex = 0;
+		}
+
+		if (assetPublisherQueryName.equals("assetCategories")) {
+			Map<Long, List<AssetCategoryModel>> assetCategoryModelsMap =
+				assetCategoryModelsMaps[(int)groupId - 1];
+
+			List<AssetCategoryModel> assetCategoryModels =
+				assetCategoryModelsMap.get(
+					_getNextAssetClassNameId(groupId, assetClassNameIds));
+
+			if ((assetCategoryModels == null) ||
+				assetCategoryModels.isEmpty()) {
+
+				return _newPortletPreferencesModel(
+					plid, portletId, PortletConstants.DEFAULT_PREFERENCES);
+			}
+
+			objectValuePair = _getAssetPublisherAssetCategoriesQueryValues(
+				assetCategoryModels, startIndex);
+		}
+		else {
+			Map<Long, List<AssetTagModel>> assetTagModelsMap =
+				assetTagModelsMaps[(int)groupId - 1];
+
+			List<AssetTagModel> assetTagModels = assetTagModelsMap.get(
+				_getNextAssetClassNameId(groupId, assetClassNameIds));
+
+			if ((assetTagModels == null) || assetTagModels.isEmpty()) {
+				return _newPortletPreferencesModel(
+					plid, portletId, PortletConstants.DEFAULT_PREFERENCES);
+			}
+
+			objectValuePair = _getAssetPublisherAssetTagsQueryValues(
+				assetTagModels, startIndex);
+		}
+
+		String[] assetPublisherQueryValues = objectValuePair.getKey();
+
+		_assetPublisherQueryStartIndexes.put(
+			groupId, objectValuePair.getValue());
+
+		PortletPreferences jxPortletPreferences =
+			(PortletPreferences)
+				_defaultAssetPublisherPortletPreferencesImpl.clone();
+
+		jxPortletPreferences.setValue("queryAndOperator0", "false");
+		jxPortletPreferences.setValue("queryContains0", "true");
+		jxPortletPreferences.setValue("queryName0", assetPublisherQueryName);
+		jxPortletPreferences.setValues(
+			"queryValues0",
+			new String[] {
+				assetPublisherQueryValues[0], assetPublisherQueryValues[1],
+				assetPublisherQueryValues[2]
+			});
+		jxPortletPreferences.setValue("queryAndOperator1", "false");
+		jxPortletPreferences.setValue("queryContains1", "false");
+		jxPortletPreferences.setValue("queryName1", assetPublisherQueryName);
+		jxPortletPreferences.setValue(
+			"queryValues1", assetPublisherQueryValues[3]);
+
+		return _newPortletPreferencesModel(
+			plid, portletId,
+			_portletPreferencesFactory.toXML(jxPortletPreferences));
+	}
+
+	public PortletPreferencesModel newPortletPreferencesModel(
+			long plid, String portletId, DDLRecordSetModel ddlRecordSetModel)
+		throws Exception {
+
+		PortletPreferences jxPortletPreferences = new PortletPreferencesImpl();
+
+		jxPortletPreferences.setValue("editable", "true");
+		jxPortletPreferences.setValue(
+			"recordSetId", String.valueOf(ddlRecordSetModel.getRecordSetId()));
+		jxPortletPreferences.setValue("spreadsheet", "false");
+
+		return _newPortletPreferencesModel(
+			plid, portletId,
+			_portletPreferencesFactory.toXML(jxPortletPreferences));
+	}
+
+	public PortletPreferencesModel newPortletPreferencesModel(
+			long plid, String portletId,
+			JournalArticleResourceModel journalArticleResourceModel)
+		throws Exception {
+
+		PortletPreferences jxPortletPreferences = new PortletPreferencesImpl();
+
+		jxPortletPreferences.setValue(
+			"articleId", journalArticleResourceModel.getArticleId());
+		jxPortletPreferences.setValue(
+			"groupId",
+			String.valueOf(journalArticleResourceModel.getGroupId()));
+
+		return _newPortletPreferencesModel(
+			plid, portletId,
+			_portletPreferencesFactory.toXML(jxPortletPreferences));
+	}
+
+	private PortletPreferenceDataFactory() {
+		_defaultAssetPublisherPortletPreferencesImpl =
+			(PortletPreferencesImpl)_portletPreferencesFactory.fromDefaultXML(
+				readFile("default_asset_publisher_preference.xml"));
+	}
+
+	private ObjectValuePair<String[], Integer>
+		_getAssetPublisherAssetCategoriesQueryValues(
+			List<AssetCategoryModel> assetCategoryModels, int index) {
+
+		String[] categoryIds = new String[4];
+
+		for (int i = 0; i < 4; i++) {
+			if (i > 0) {
+				index +=
+					BenchmarksPropsValues.
+						MAX_ASSET_ENTRY_TO_ASSET_CATEGORY_COUNT;
+			}
+
+			AssetCategoryModel assetCategoryModel = assetCategoryModels.get(
+				index % assetCategoryModels.size());
+
+			categoryIds[i] = String.valueOf(assetCategoryModel.getCategoryId());
+		}
+
+		return new ObjectValuePair<>(
+			categoryIds,
+			index +
+				BenchmarksPropsValues.MAX_ASSET_ENTRY_TO_ASSET_CATEGORY_COUNT);
+	}
+
+	private ObjectValuePair<String[], Integer>
+		_getAssetPublisherAssetTagsQueryValues(
+			List<AssetTagModel> assetTagModels, int index) {
+
+		String[] assetTagNames = new String[4];
+
+		for (int i = 0; i < 4; i++) {
+			if (i > 0) {
+				index +=
+					BenchmarksPropsValues.MAX_ASSET_ENTRY_TO_ASSET_TAG_COUNT;
+			}
+
+			AssetTagModel assetTagModel = assetTagModels.get(
+				index % assetTagModels.size());
+
+			assetTagNames[i] = String.valueOf(assetTagModel.getName());
+		}
+
+		return new ObjectValuePair<>(
+			assetTagNames,
+			index + BenchmarksPropsValues.MAX_ASSET_ENTRY_TO_ASSET_TAG_COUNT);
+	}
+
+	private long _getNextAssetClassNameId(
+		long groupId, long[] assetClassNameIds) {
+
+		Integer index = _assetClassNameIdsIndexes.get(groupId);
+
+		if (index == null) {
+			index = 0;
+		}
+
+		long classNameId = assetClassNameIds[index % assetClassNameIds.length];
+
+		_assetClassNameIdsIndexes.put(groupId, ++index);
+
+		return classNameId;
+	}
+
+	private PortletPreferencesModel _newPortletPreferencesModel(
+		long plid, String portletId, String preferences) {
+
+		PortletPreferencesModel portletPreferencesModel =
+			new PortletPreferencesModelImpl();
+
+		// PK fields
+
+		portletPreferencesModel.setPortletPreferencesId(counter.get());
+
+		// Audit fields
+
+		portletPreferencesModel.setCompanyId(COMPANY_ID);
+
+		// Other fields
+
+		portletPreferencesModel.setOwnerId(PortletKeys.PREFS_OWNER_ID_DEFAULT);
+		portletPreferencesModel.setOwnerType(
+			PortletKeys.PREFS_OWNER_TYPE_LAYOUT);
+		portletPreferencesModel.setPlid(plid);
+		portletPreferencesModel.setPortletId(portletId);
+		portletPreferencesModel.setPreferences(preferences);
+
+		return portletPreferencesModel;
+	}
+
+	private static PortletPreferenceDataFactory _portletPreferenceDataFactory =
+		new PortletPreferenceDataFactory();
+
+	private final Map<Long, Integer> _assetClassNameIdsIndexes =
+		new HashMap<>();
+	private final Map<Long, Integer> _assetPublisherQueryStartIndexes =
+		new HashMap<>();
+	private final PortletPreferencesImpl
+		_defaultAssetPublisherPortletPreferencesImpl;
+	private final PortletPreferencesFactory _portletPreferencesFactory =
+		new PortletPreferencesFactoryImpl();
+
+}
