+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/macro.ftl
@@ -1,7 +1,8 @@
 <#setting number_format = "computer">
 
 <#assign
-	classNameDataFactory = dataFactory.getDataFactoryInstance()
+	classNameDataFactory = dataFactory.getDataFactoryInstance("classNameDataFactory")
+	portletPreferenceDataFactory = dataFactory.getDataFactoryInstance("portletPreferenceDataFactory")
 />
 
 <#macro insertAssetEntry
